+++ CHANGELOG.md
@@ -8,6 +8,15 @@ Pre-release
 -----------
 
 Language changes:
+ * Line types. More exp/log types with steepness not 6.
+    - Add `exp11`, `log11`, `xpe11`, `lge11`. Steepness 11.
+
+v0.5.1 (2025-02-16)
+-------------------
+
+Add ADSR envelope to sweepable parameters.
+
+Language changes:
  * Parameter envelopes. Implement ADSR envelope, add to every
    sweepable parameter. This uses list heading subparameters,
    under the new `.e` (envelope) subparameter which has a new